refactor(invoice): move to script setup
